Properly format date/time values
Since timezones matter for us (we want local times at the destination, not
times in our current timezone), we can't pass date/time values through QML,
as the timezone gets lost in the conversion to JS Date. So we need to do
the property read as part as the formatting function.