Force OpenSSL into the APK
QtNetwork needs it, but it doesn't link against it but dlopens it at
runtime. This makes the weather forecast work on Android too.

See http://doc.qt.io/qt-5/opensslsupport.html for mostly working OpenSSL
build instructions on Android. I needed at least API level 21 for this to
work, as well as OpenSSL 1.0.x for the pre-built Qt 5.10.