Improve the reservation number based trip group search
We now allow multiple reservation numbers for the same mode of transport
if they are connected by adjacent transitions in a location that is not
the starting point.

This is a bit too aggressive still, but overall seems to produce more
reliable results on non-trivial trips than the previous approach.