Initial work on public transport data online access
This just contains barely enough data model and backend access code to
search for a journey between two locations. Most notably this still misses
a frontend API that hides the various low-level REST calls behind a
job-like interface.

This is prepared to eventually become a standalone framework, it's useful
beyond just this application.