Implement onNewIntent
Summary: Importing a reservation from an intent only worked when that intent started the app. When the app is already running onNewIntent is called, but since we didn't override it nothing was imported.

Test Plan: Open app. Send reservation via KDE Connect. Previously nothing happend. Now the reservation is added to the list

Reviewers: vkrause

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D17993