Load cached live data before checking for reservation relevancy
This matters if you restart the application while you are on a delayed
train for example, after the scheduled arrival time. As we loaded the
cached delay data afterwards, the corresponding trip was wrongly assumed
to already have arrived (as we could only compare to the base schedule),
and therefore got discarded for further live querying.