Support runtime permissions
Summary:
To be able to release Itinerary in Google Play we need to target SDK version 28. This requires us to support the permission system introduced in Marshmallow. Before we access the calendar we need to check and potentially request
the calendar permission.

It seems like we don't actually need the storage permission, we get access rights to files received via intents automatically.

Reviewers: vkrause

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D18409