Rework the place delegate
Most of the logic on which actions to show is now in there, and supported
via the timeline delegate controller. This makes things consistent between
the timeline delegates and the details pages. It also enables access to
more details in the place delegate, so we can now show departure
information with the correct times.