Add a wrapper class and macro for static JNI properties
This avoids boilerplate code, and allows us to use properties as such
rather than representing them as functions. Read-only so far, as we only
use this for constants, but the concept should work for writable and
non-static properties as well, if we ever need that.

This however requires C++17 inline variables, which isn't a problem as we
only need this on Android, and there the compiler is new enough.