Add a basic development mode
It's activated Android-style by tapping the version number in the about
screen 7 times, and can be deactivated again on a dedicated page.

So far this activates all of the debug features and diagnostics that were
already there bug controlled by comments. But there's more to come to make
MapCSS editable easily on Android too.