Allow to enable full KPublicTransport logging in development mode
This is particularly useful on Android, where we can't just set the
environment variable externally, and where we have to deal with getting
the logs out of the app containment.

While at it, this also removes the use of hard-coded permission strings
from other places in the code in favor of JNI property access.