Split out the Java part of SolidExtras as well
This avoids entangling app code and code that eventually should move to
KF5 further here, in particular for the upcoming network status work.

This requires making SolidExtras a dynamic library on Android for its AAR
to be picked up by androiddeployqt. For this to work without requiring a
make install first, a few ECM changes are needed though.