Add APK exclusion lists for libraries and assets
This removes a number of plugins and libraries that androiddeployqt adds
due to some rather broad dependency rules, but that we don't actually need.

This also reverts parts of the broad inclusion patterns applied by ECM,
which results in a lots of extra assets from share/ that we don't need
or want either.

For the latter we can ideally fix more of that at the source itself, but
for now that's a good stop-gap measure.