Set the APK version based on what is defined in CMake
Most of the existing solutions we have do this by patching the manifest in
the source directory, due to what androiddeployqt expects. This approach
uses Gradle's built-in mechanisms instead, which basically gives us a
hook into the middle of androiddeployqt's processing. To pass values from
CMake to Gradle we use a file we can include in the build.gradle file and
that we place next to androiddeployqt's working directory.

This also does the automatic version code bumping via Gradle that is
currently done by the nightly build script on binary factory, as we need
the same for all other builds as well.