Force using org.kde.desktop style on non-Android
On Plasma this is done via plasma-integration, but on non-Plasma systems
we get the default QML style that is severly broken.

Set the style to org.kde.desktop but allow to override it via env var so
we can get org.kde.breeze on Plasma Mobile.

Also mark it as a dependency in CMake