Add a command line option for importing temporary files
Due to being a unique application, we cannot just receive temporary content
(e.g. from Qrca) via stdin, nor can the calling application properly manage
the lifetime of a temporary file for that, as the called process can return
before an already running instance was able to process to data.

Instead we now have an explicit option to tell the consuming instance to
clean up the files it is being passed once those have been imported.