Fix ReservationHelper::equals always returning false
We miss explicit equal comparator registration for types held as variants
inside Reservation objects. Ideally this is something done in the library
eventually, but it's not straightforward to do early enough there, in
particular in static builds. So we do this here for now.

This fixes live data updates triggering reservation changes
unconditionally, and thus everything that follows from there.