[Delegates] Use preferredWidth instead of width
When an item is inside a Layout, its width (and height) gets managed by said
layout, taking into account minimum/maximum/preferredWidth and implicitWidth.

Assigning width manually, does not behave as one expect here, since
Kirigami.Icon has an implicit size (of 32) and so with the nature of a ListView
creating delegates as you scroll, it becomes somewhat random which size
wins in the end, and might even change as an item goes out of view and back in.

While at it, use `smallMedium`, which, from the actual looks of the outcome
of this bug, is what the delegate should have been using. (Imho small is too small)