Add line or product logos to the timeline delegate header when available
As those can be non-square that requires proper aspect ratio handling.
Unfortunately Kirigami.Icon hardcodes its implicit size to 32x32, giving
us no way to do this properly. This problem exists in numerous other places
where we use those icons as well, those just happen to be a bit smaller
in height, so the implicitWidth of 32 just happens to work ok-ish.

For the timeline headers this looks broken though, so we need a correct
solution, even if the implementation is ugly. For this we now have a helper
method that peaks into the SVG data to read out the view box size and
computes the aspect ratio from that.

Still needs to be rolled out to all the other places using line icons.