Allow to shorten train trips
That is, for an existing reservation for which we know the intermediate
stops, allow to change the departure to a later stop and the arrival to
an earlier one. This is useful when having to react to disruptions or
having to do creative bookings for discounts, and is more convenient
compared to manually editing arrival/departures (and which we can't even
do yet at all anyway).