Ensure QAndroidJniObjects from implicit conversion outlive the JNI call
When passing in complex types that we can implicitly convert we went
directly to the jobject handle via a temporary QAndroidJniObject so far,
which at the point of the actual call has been invalid already. Instead
we now split argument processing into two parts. In the first step we only
go up to the QAndroidJniObject, and we store those on the stack to outlive
the JNI call. The second step then does the final conversion to the jobject
handle.