Let ApplicationController control pkpass imports entirely
PkPassManager driving this itself is a leftover from the very very early
days of this.

As an immediate benefit this will now also import pkpass files in a deeper
nested structure such as as in email attachments. This also clears the
way for supporting passes that are not tied to a timeline element.