Don't make JNI_OBJECTs implicitly constructable from JNI handles
This is the equivalent of reinterpret_cast-ing from a void*, so this should
be avoided and used only when absolutely necessary. Add a helper method
for doing this.

This also makes single argument ctors with wrapper type arguments work as
intended, those were made ambigious by the ctor added via JNI_OBJECT.