Check availability of documents while counting them
Fixes the cases where documents are counted but not available.
This can happen if the events come from android's system calendar or if the document was deleted from another event's document list.
In both of these cases, the reference should probably be deleted instead, but this is a good safety measure nevertheless.

BUG: 456510