Move driving side label computation to C++
The driving side enum is a uint8_t, which QML apparently doesn't like and
we thus end up with arbitrary values there. As the "driving side changed"
flag is correct, and due to the way the driving side enum was queried we
ended up with "left" for any bogus value. That is however only wrong if
your native driving side is left as well.

FIXED-IN: 22.12.0
BUG: 461438