Work around KUnitConversion deadlocking when needing a currency update
The download of the exchange rate data is supposed to be blocking, but
that doesn't really work with QML apparently, and we can re-enter this
method a second time, which then deadlocks on the call into
KUnitConversion.

This needs a proper fix in KUnitConversion in KF6.

BUG: 469562
FIXED-IN: 23.04.2

(cherry picked from commit 5a2971fc4bacddc41a2fc895640aee61bebb3157)