Avoid including the address country if it's obvious
Introduces Localizer.formatAddressWithContext which can then decide
whether to include the country when formatting the address (and/or
other criteria in the future).

The country is excluded if the address is otherwise empty and in the
user's home country. If there's an address pair (e.g. departure and arrival)
the country is hidden only if both are in the same country.