Work around the Android back key closing the app on secondary layers
Secondary layers on Android don't seem to get focus by default, which
results in the back key event falling through to the default handling
in Qt, which closes the app.

This is highly undesired and happens for secondary pages that were never
interacted with (and thus never got focus). To work around this we now
force focus on the top layer whenever that changes.