Fail less dramatically when ending up with an unknown pass
Happens e.g. on corrupt storage (e.g. when switching from master to
23.04 at the moment). Returning a pass type that the DelegateChooser
doesn't handle results in an entirely empty view. Returning a valid
type in that case will still result in a broken entry, but an otherwise
functional view that allows you to delete the broken entry and thus
recover from the issue.