Ensure we pass content: URLs fully encoded to Android API and QFile
This fixes the app crashing due to an unhandled security exception when
trying to open a file with e.g. a space anywhere in its name or path.

content: URLs we get from the file dialog need to get back into Android
API (via QFile or otherwise) as exactly the same string, any kind of
changes e.g. to the URL encoding are not considered to be the same file,
and thus a file we have no permission to access to.

BUG: 470643
FIXED-IN: 23.04.3