Handle online ticket retrieval URLs
Those cannot be GET'ed, but are intended for the corresponding apps to
get the necessary parameters for API calls to retrieve the ticket. So
parse those and pass on the relevant arguments to the online ticket
retrieval job.

This is still hardcoded just for DB, we still need a declarative way to
describe online ticket access.

With this working we could also register as URL handlers for those URLs
on Android, so Itinerary can be opened automatically by clicking on such
a URL in an email for example.