Convert QML for Qt 6/KF6 compatibility when building for Qt 6
This is an alternative to the commonly used CMake preprocessing approach.
It's slightly more flexible and could potentially be extended for
supporting inline alternative code blocks for more complex changes.

The main advantage over the CMake approach is that the QML in the source
dir remains valid and can be consumed by qmlimportscanner/androiddeployqt,
and thus not breaking APKs for Qt 5.

The price we pay for that are in-place changes in the source dir, which
is rather ugly.

The actual changes included so far are basically what is in !171 and !173,
but on their own are not yet sufficient to start up on Qt 6. In particular
the Kirigami page actions and the barcode scanner still need work.