Centralize QNetworkAccessManager creation
Instead of doing that in a bunch of classes pass around a factory
function to a central location. This is needed to still allow on
demand creation of the QNAM instance, as that is somewhat expensive
and we don't want it on the startup path.

Side-effect of this is that we can test these classes easier using
the mock QNAM.

Still to be done for the other QNAM users though.