Add method to match and cut out a sub-journey from a full train run
This is a necessary building block for integrating data from onboard
train API. We'll use that to find the part relevant for the current
reservation.

This is relatively strict for what is considered a matching stop,
which we need due to the varying data quality of some of the onboard
APIs.