Rework the rest of the transport icon uses
The way non-square SVG files are displayed changed, in a way not suitable
for us here. Use Image rather than Icon for the line logos instead, to
work around that. That fixes the sizes but still gives us a worse render
quality for SVGs than before (possibly due to rendering the SVG in its
original size and then downscaling rather than rendering in the needed
size to begin with). But at least we now have a central place to address
that going forward.