Pick a better default text color for pkass event tickets
This matters for passes that contain a background image contrary to the
default background of the current palette and that don't specify text
colors themselves. This is often observed in Pretix passes, with the 37C3
being particularly bad having a mostly black background image.

In that case we now decide the default text color based on the background
image content. Kirigami.ImageColors would seem like the perfect tool for
that, but that unfortunately fails for black/white content, so we need to
do this ourselves here, using a primitive equidistant sampling approach.