Don't add file type filters for the input file dialog on Android
As there generally isn't a file type selector on Android, if multiple
filters are specified we get the superset of all all of them with Qt 6,
but excluding the "all files" one. This is too restrictive for us here,
as the list is far from complete (and hard to complete due to the various
image file extension variations).