Handle intermediate stops without coordinates
Previously, the path was just not drawn.
Now the station with missing data is just skipped.

This can be tested with the Vilnius to Riga connection (make sure to set
the departure time to early in the morning, the backend doesn't find
connections on the following day yet, and the train goes early and only
once a day)