Don't compute trip group start/end times on demand
Instead, store them in the TripGroup object and update them when needed.

This implements tracking, computation and de/serialization, as well as
migrating legacy data. It doesn't actually make use of this yet though,
in order to first validate the results against the old code.

This also fixes an infinite update loop on transfer changes.