Fix seat label text overrun
This is particularly prone to happen when having many seats reserved in
a multi-traveler scenario.

We now try to keep equally sized labels until that would require eliding
a single label, and then prefer unequally sized labels over eliding. If
there's still not enough space for that, we elide the content.