Determine APK version code from CI_PIPELINE_CREATED_AT when set
This ensures all APKs build in the same pipeline (ie. those for the
different architectures) have the same version code. This is necessary
as both F-Droid and Google Play don't like 32bit APKs with higher version
codes than the corresponding 64bit one.

Local builds still work as before, and we can still mix local and CD
builds for development without having to go down in version codes.

Adapt variable naming for forward-compatibility with KF 6.12 which provides
the equivalent of version.gradle out of the box.