Unify temperature range formatting
Compare rounded values in the target unit to decide whether we show a
single value or two.

Also, make the decision to use Fahrenheit a property binding on Settings,
which results in units being automatically updated when changing the
country.

This also prepares us for a possible "force normal units" setting which
we'll likely need when switching to the new distance formatting API from
KCoreAddons which supports imperial units.