Work around pkpass type identifier case-sensitivity getting lost in URLs
We have paths that pass this through a QUrl::host, which converts to
lower-case, and are then unable to find the correct pkpass file on disk
again.

This shouldn't happen in the first place, but since it's part of the stored
schema.org data we have to have a fallback either way to recover already
existing data affected by this.

CCBUG: 505891