Distinguish between editing manual and KPublicTransport-backed boat trips
Manual ones allow free entry of departure/arrival locations and times,
while those backed by schedule data only allow to select stops on the
schedule (like we currently do for trains/busses) already.

If this proves to be useful this approach could be applied to all other
modes of transportation as well.