Don't apply ticket issuer filters for journey queries with coordinates
The coverage metadata is meanwhile good enough for coordinates to result
in the better and more reliable results on their own, the ticket issuer
is unnecessarily restricting there.

One noticeable difference with this is that Transitous gets considered
much more often for alternative connection searches, which is particularly
useful when the originally booked operator is entirely out of service,
e.g. giving you long distance bus options as an alternative to a train.