Write more detailed version information to export files
For CD builds this includes a Git commit hash, making it easier to
identify how a specific export was actually created.

For regular use this doesn't matter, but exports are also frequently
shared for providing sample data or test cases for issues, and for that
the exact producer version is super helpful.