Query nominatim to find railway station coordinates in reservations
This allows to select a correct backend even if there is no matching API
for the ticket seller.

The alternatives page is now disabled until the coordinates are
resolved, to avoid querying all APIs. If resolving the stations
failed, you can help the geocoder by specifying the country, city etc.