Load full reservation batches and merge their common parts
This makes sure that the common parts all have the same information,
which fixes a number of issues in multi-ticket scenarios where e.g.
realtime data retrieval or map display doesn't work correctly because
it's working with one of the reservations in the batch without the full
information.

Cost-wise we can do that now that we don't load all reservation batches
at startup anymore, but just those actually displayed or otherwise used.