Correctly check for invalidated sort order on reservation changes
We need to use the same criteria as for sorting itself, otherwise we can
end up in an inconsistent state. This could happen here for multiple
distinct reservations with the same start but a different end time.

This fixes a trip disassociation if one such reservation is changed and
happens to be the first in a trip group.